+++ context/app/static/js/components/Header/DropdownLink/index.js
@@ -1,3 +1,0 @@
-import DropdownLink from './DropdownLink';
-
-export default DropdownLink;