+++ context/app/static/js/components/Header/Header/Header.tsx
@@ -1,4 +1,4 @@
-import React, { useRef } from 'react';
+import React from 'react';
 
 import EntityHeader from 'js/components/detailPage/entityHeader/EntityHeader';
 import HeaderAppBar from '../HeaderAppBar';
@@ -6,13 +6,12 @@ import HeaderContent from '../HeaderContent';
 import { useEntityHeaderVisibility } from './hooks';
 
 function Header() {
-  const anchorRef = useRef<HTMLDivElement>(null);
   const { shouldDisplayHeader, ...props } = useEntityHeaderVisibility();
 
   return (
     <>
-      <HeaderAppBar anchorRef={anchorRef} {...props}>
-        <HeaderContent anchorRef={anchorRef} />
+      <HeaderAppBar {...props}>
+        <HeaderContent />
       </HeaderAppBar>
       {shouldDisplayHeader && <EntityHeader />}
     </>