+++ context/app/static/js/components/detailPage/BulkDataTransfer/index.js
@@ -1,3 +1,0 @@
-import BulkDataTransfer from './BulkDataTransfer';
-
-export default BulkDataTransfer;