+++ context/app/static/js/components/detailPage/entityHeader/EntityHeaderItem/utils.js
@@ -1,6 +1,0 @@
-export const truncateText = (inputText) => {
-  if (inputText.length > 100) {
-    return `${inputText.slice(0, 100)}...`;
-  }
-  return inputText;
-};