+++ context/app/static/js/components/types.ts
@@ -1,3 +1,6 @@
+import { ContributorAPIResponse } from './detailPage/ContributorsTable/utils';
+import { UnprocessedFile } from './detailPage/files/types';
+
 export type DonorEntityType = 'Donor';
 export type SampleEntityType = 'Sample';
 export type DatasetEntityType = 'Dataset';
@@ -23,14 +26,25 @@ export type DagProvenanceType =
 
 export interface Entity {
   entity_type: ESEntityType;
+  description: string;
   uuid: string;
   hubmap_id: string;
   last_modified_timestamp: number;
   created_timestamp: number;
   ancestors: Entity[];
+  // eslint-disable-next-line no-use-before-define -- Donor is defined later in the file and extends Entity
+  donor: Donor;
   descendant_counts: { entity_type: Record<string, number> };
+  group_name: string;
+  created_by_user_displayname: string;
+  created_by_user_email: string;
+  mapped_status: string;
+  mapped_data_access_level: 'Public' | 'Protected' | 'Consortium';
+  status: string;
   [key: string]: unknown;
 }
+
+export type PartialEntity = Partial<Entity> & Pick<Entity, 'entity_type' | 'uuid' | 'hubmap_id'>;
 
 export interface Donor extends Entity {
   entity_type: 'Donor';
@@ -44,11 +58,12 @@ export interface Donor extends Entity {
 
 export interface Sample extends Entity {
   entity_type: 'Sample';
-  sample_category?: string;
+  sample_category: string;
   mapped_organ: string;
   organ: string;
   metadata?: Record<string, string>;
   origin_samples_unique_mapped_organs: string[];
+  origin_samples: Sample[];
 }
 
 export interface Dataset extends Entity {
@@ -69,6 +84,16 @@ export interface Dataset extends Entity {
   thumbnail_file?: {
     file_uuid: string;
   };
+  dbgap_study_url: string;
+  dbgap_sra_experiment_url: string;
+  files: UnprocessedFile[];
+  contributors: ContributorAPIResponse[];
+  sub_status: string;
+  protocol_url: string;
+  registered_doi: string;
+  doi_url: string;
+  published_timestamp: number;
+  mapped_external_group_name?: string;
 }
 
 export interface Collection extends Entity {
@@ -81,32 +106,42 @@ export interface Publication extends Entity {
 
 export interface Support extends Entity {
   entity_type: 'Support';
+  origin_samples: Sample[];
+  files: UnprocessedFile[];
+  contributors: ContributorAPIResponse[];
+  published_timestamp: number;
+  assay_modality: 'single' | 'multiple';
+  created_timestamp: number;
+  last_modified_timestamp: number;
+  mapped_data_types: string[];
 }
 
-function checkEntityType(entity: Entity, entityType: ESEntityType): boolean {
+export type EntityWithType = Partial<Entity> & Required<Pick<Entity, 'entity_type'>>;
+
+function checkEntityType(entity: EntityWithType, entityType: ESEntityType): boolean {
   return entity.entity_type === entityType;
 }
 
-export function isDataset(entity: unknown): entity is Dataset {
-  return checkEntityType(entity as Entity, 'Dataset');
+export function isDataset(entity: EntityWithType): entity is Dataset {
+  return checkEntityType(entity, 'Dataset');
 }
 
-export function isSample(entity: unknown): entity is Sample {
-  return checkEntityType(entity as Entity, 'Sample');
+export function isSample(entity: EntityWithType): entity is Sample {
+  return checkEntityType(entity, 'Sample');
 }
 
-export function isDonor(entity: unknown): entity is Donor {
-  return checkEntityType(entity as Entity, 'Donor');
+export function isDonor(entity: EntityWithType): entity is Donor {
+  return checkEntityType(entity, 'Donor');
 }
 
-export function isCollection(entity: unknown): entity is Collection {
-  return checkEntityType(entity as Entity, 'Collection');
+export function isCollection(entity: EntityWithType): entity is Collection {
+  return checkEntityType(entity, 'Collection');
 }
 
-export function isPublication(entity: unknown): entity is Publication {
-  return checkEntityType(entity as Entity, 'Publication');
+export function isPublication(entity: EntityWithType): entity is Publication {
+  return checkEntityType(entity, 'Publication');
 }
 
-export function isSupport(entity: unknown): entity is Support {
-  return checkEntityType(entity as Entity, 'Support');
+export function isSupport(entity: EntityWithType): entity is Support {
+  return checkEntityType(entity, 'Support');
 }